Begin breaking out cli_test.py
